# [quantization] Vectorized specialization of max_pool2d for channels-last layout