--- a/aten/src/ATen/native/quantized/cpu/kernels/QuantizedOpKernels.cpp
+++ b/aten/src/ATen/native/quantized/cpu/kernels/QuantizedOpKernels.cpp
@@ -117,12 +117,110 @@ void qadd_kernel(Tensor& out, const Tensor& self, const Tensor& other) {
   });
 }
 
+void qmaxpool_2d_nhwc_kernel(const Tensor &qx,
+                             int64_t iC, // input/output channels
+                             int64_t iH,
+                             int64_t iW, // input sizes
+                             int64_t oH,
+                             int64_t oW, // output sizes
+                             int64_t kH,
+                             int64_t kW, // kernel size
+                             int64_t sH,
+                             int64_t sW, // strides
+                             int64_t pH,
+                             int64_t pW, // padding
+                             int64_t dH,
+                             int64_t dW, // dilation
+                             Tensor &qy) {
+  AT_DISPATCH_QINT_TYPES(qx.scalar_type(), "max_pool2d_nhwc", [&]() {
+    scalar_t *idata = qx.data<scalar_t>();
+    scalar_t *odata = qy.data<scalar_t>();
+
+    // Loop over N
+    for (int64_t b = 0; b < qx.size(0); ++b) {
+      // Loop over H
+      auto *i_p = reinterpret_cast<typename scalar_t::underlying*>(idata + b * iW * iH * iC);
+      for (int64_t row = 0; row < oH; ++row) {
+        // Loop over W
+        for (int64_t col = 0; col < oW; ++col) {
+          // Pointer to output data for this specific N,H,W position
+          auto *o_p = reinterpret_cast<typename scalar_t::underlying*>(odata + b * oH * oW * iC + row * oW * iC + col * iC);
+
+          // Loop over reduction block
+          int64_t h_start = row * sH - pH;
+          int64_t w_start = col * sW - pW;
+          int64_t h_end = std::min(h_start + (kH - 1) * dH + 1, iH);
+          int64_t w_end = std::min(w_start + (kW - 1) * dW + 1, iW);
+          while (h_start < 0)
+            h_start += dH;
+          while (w_start < 0)
+            w_start += dW;
+
+          int64_t c = 0;
+
+          // Interleaved vector loop 4x
+          constexpr auto vec_width = Vec256<scalar_t>::size();
+          for (; c + 4 * vec_width <= iC; c+= 4 * vec_width) {
+            Vec256<scalar_t> acc(scalar_t(std::numeric_limits<typename scalar_t::underlying>::lowest()));
+            Vec256<scalar_t> accs[4] = {acc, acc, acc, acc};
+            int64_t tcntr = 0;
+            int64_t x, y;
+            for (y = h_start; y < h_end; y += dH) {
+              for (x = w_start; x < w_end; x += dW) {
+                for (int i = 0; i < 4; ++i) {
+                  tcntr = y * iW + x;
+                  auto vals = Vec256<scalar_t>::loadu(i_p + tcntr * iC + c + Vec256<scalar_t>::size() * i);
+                  accs[i] = vec256::maximum(accs[i], vals);
+                }
+              }  // for x
+            }  // for y 
+            for (int i = 0; i < 4; ++i) {
+              accs[i].store(o_p + c + Vec256<scalar_t>::size() * i);
+            }
+          }  // for c
+
+          // Vector loop
+          for (; c + vec_width <= iC; c+= vec_width) {
+            Vec256<scalar_t> acc(scalar_t(std::numeric_limits<typename scalar_t::underlying>::lowest()));
+            int64_t tcntr = 0;
+            int64_t x, y;
+            for (y = h_start; y < h_end; y += dH) {
+              for (x = w_start; x < w_end; x += dW) {
+                tcntr = y * iW + x;
+                auto vals = Vec256<scalar_t>::loadu(i_p + tcntr * iC + c);
+                acc = vec256::maximum(acc, vals);
+              }  // for x
+            }  // for y 
+            acc.store(o_p + c);
+          }  // for c
+
+          for (; c < iC; ++c) {
+            auto max_val = std::numeric_limits<typename scalar_t::underlying>::lowest();
+            int64_t tcntr = 0;
+            int64_t x, y;
+            for (y = h_start; y < h_end; y += dH) {
+              for (x = w_start; x < w_end; x += dW) {
+                tcntr = y * iW + x;
+                auto val = *(i_p + tcntr * iC + c);
+                max_val = std::max(max_val, val);
+              }  // for x
+            }  // for y
+
+            o_p[c] = max_val;
+          }  // for c
+        }  // for col
+      }  // for row
+    }  // for b
+  });
+} 
+
 } // namespace
 
 REGISTER_DISPATCH(qrelu_stub, &qrelu_kernel);
 REGISTER_DISPATCH(qrelu6_stub, &qrelu6_kernel);
 REGISTER_DISPATCH(qadd_relu_stub, &qadd_kernel<true>);
 REGISTER_DISPATCH(qadd_stub, &qadd_kernel<false>);
+REGISTER_DISPATCH(qmaxpool_2d_nhwc_stub, &qmaxpool_2d_nhwc_kernel);
 
 } // namespace native
 } // namespace at